+++ R/framework_ebp.R
@@ -7,16 +7,17 @@
 
 
 framework_ebp <- function(fixed, pop_data, pop_domains, smp_data, smp_domains, 
-                     threshold, custom_indicator = NULL, na.rm, weights, smp_weights = NULL,
+                     threshold, custom_indicator = NULL, na.rm, weights = NULL,
                      pop_weights = NULL) {
 
   # Reduction of number of variables
   mod_vars <- all.vars(fixed)
   mod_vars <- mod_vars[mod_vars != as.character(fixed[2])]
   smp_vars <- c(as.character(fixed[2]), mod_vars, smp_domains, weights)
-  pop_vars <- c(mod_vars, pop_domains)
+  pop_vars <- c(mod_vars, pop_domains, pop_weights)
   smp_data <- smp_data[, smp_vars]
   weights  <- weights
+  pop_weights <- pop_weights
   fw_check1(pop_data = pop_data, mod_vars = mod_vars, pop_domains = pop_domains, 
            smp_data = smp_data, fixed = fixed, smp_domains = smp_domains, 
            threshold = threshold, weights = weights)
@@ -24,7 +25,6 @@ framework_ebp <- function(fixed, pop_data, pop_domains, smp_data, smp_domains,
 
   pop_data <- pop_data[, pop_vars]
 
-  
   # Deletion of NA
   if (na.rm == TRUE) {
     pop_data <- na.omit(pop_data)
@@ -34,23 +34,13 @@ framework_ebp <- function(fixed, pop_data, pop_domains, smp_data, smp_domains,
           ebp.')
   }
   
-  if (is.null(pop_weights) == TRUE){
+
+  if (is.null(pop_weights) == FALSE & is.null(weights) == TRUE){
     
-    popweight_data <- rep(1, nrow(pop_data))
-     
-  }
-  
-  popweight_data <- pop_data[,pop_weights]  
-  pop_data <- pop_data[, pop_vars]
-  
-  if(is.null(smp_weights) == TRUE){
-    
-    smpweight_data <- rep(1, nrow(smp_data))
+    message('Missing sample weights, vector of 1s created')
+    pop_data$weights <- rep(1, nrow(smp_data))
     
   }
-  
-  smpweight_data <- smp_data[,smp_weights]
-  smp_data <- smp_data[,c(smp_vars,smp_weights)]
   
   
   # Order of domains
@@ -146,6 +136,7 @@ framework_ebp <- function(fixed, pop_data, pop_domains, smp_data, smp_domains,
           the median of the dependent variable and equals",threshold, "\n")
   }
   
+  
 
   return(list(pop_data         = pop_data,
               pop_domains_vec  = pop_domains_vec,
@@ -166,8 +157,7 @@ framework_ebp <- function(fixed, pop_data, pop_domains, smp_data, smp_domains,
               indicator_names  = indicator_names,
               threshold        = threshold,
               weights          = weights,
-              smp_data         = smp_data,
-              pop_data         = pop_data
+              pop_weights      = pop_weights
               )
          )
 }